feat: add additional password to create and delete products
As the admin username and password are exposed so anyone can see the pages only the admin has
access, I decided to add a very simple password authentication to prevent anyone from adding or
deleting products. Currently, it is not a secure method, just a thin layer of security. I may
increase the security in the future.